Do not normalize comment/mirror at 5/8 max
Previously, we were taking the max number of comments per user pair and
normalizing the comment component with that, so that the user pair with
the most number of comments were assigned 5.  For example, if A made 10
comments on B's post whereas X made 100 comments on Y's post (the
network max), lt(X,Y) would get +5 boost, and lt(A,B) would get +0.5
boost (because A->B count was only 10 compared to max 100).

This has an undesirable property where a new max suddenly dulls down
overall effectiveness.  For example, if P spammed Q with 1,000,000
comments, then the network-wide max would become 10,000x and everyone
else's comment-based component would be slashed by that factor.  This
would not be a problem if LT was solely based upon this factor, but
compared to other components such as follows, it greatly reduces its
efficacy.  For example, when someone said f1c5m15 they want comment
components to be 5x stronger than follows, no matter what the
network-wide circumstance was.

Fix this by taking the maximum out of consideration, and simply multiply
the number of comments/mirrors by the unit weight (5/8).